Fixed some timing bugs with delete, post, and update features. Also a…

+++ server/routes/surveys.router.js
@@ -8,7 +8,7 @@ const router = express.Router();
 
 router.get('/:id', (req, res) => {
   const userId = req.params.id;
-  console.log('here is the userId in the survey router', userId)
+  console.log('here is the userId in the survey router', userId);
   const queryText = `SELECT * FROM "survey_table" WHERE "user_id" = $1 ORDER BY "id" ASC;`;
   pool
     .query(queryText, [userId])
@@ -21,19 +21,18 @@ router.get('/:id', (req, res) => {
     });
 });
 
-
- 
 //Post a new survey
 router.post('/', (req, res) => {
-
   const newSurvey = req.body;
   const queryText = `INSERT INTO "survey_table" ("user_id", "survey_name")
-    values ($1, $2);`;
+    values ($1, $2)
+    RETURNING "user_id";`;
 
   pool
     .query(queryText, [newSurvey.user_id, newSurvey.survey_name])
     .then((result) => {
-      res.send(result.rows);
+      console.log('Result is:', result.rows[0]); //ID IS HERE!
+      res.send(result.rows[0]);
       console.log(`POST successful in surveys.router:`, result.rows);
     })
     .catch((err) => {
@@ -44,10 +43,10 @@ router.post('/', (req, res) => {
 
 router.put('/:id', (req, res) => {
   // Update this single title
-  console.log('Here is the req.params and req.body', req.params, req.body)
+  console.log('Here is the req.params and req.body', req.params, req.body);
   const id = req.params.id;
   // const questions = req.body.questions;
-  
+
   const queryText = `UPDATE survey_table SET survey_name = $1 WHERE id = $2`;
   pool
     .query(queryText, [req.body.survey_name, id])
@@ -62,13 +61,14 @@ router.put('/:id', (req, res) => {
 
 router.delete('/:id', (req, res) => {
   // Delete this survey
-  console.log('Here is the req.params.id', req.params.id)
+  console.log('Here is the req.params.id', req.params.id);
   const idToDelete = req.params.id;
-  const queryText = `DELETE FROM survey_table WHERE id = $1`;
-  pool 
+  const queryText = `DELETE FROM survey_table WHERE id = $1 RETURNING "user_id";`;
+  pool
     .query(queryText, [idToDelete])
     .then((result) => {
-      res.sendStatus(200);
+       console.log('survey:Delete Result is:', result.rows[0]);
+       res.send(result.rows[0]);
     })
     .catch((error) => {
       console.log(`Error making database query ${queryText}`, error);
